Update patch set 4
Patch Set 4:

Are you sure this actually makes the loudspeaker function correctly? I pushed the .xml file to my Z3 (the mixer_path.xml is identical for the Z2 and Z3), and while the dialer would indicate that the loudspeaker was enabled, I didnt get any audio signals through.

Patch-set: 4